+++ easypost/customs_info.py
@@ -1,0 +1,5 @@
+from easypost.resource import CreateResource
+
+
+class CustomsInfo(CreateResource):
+    pass