+++ easypost/customs_item.py
@@ -1,0 +1,5 @@
+from easypost.resource import CreateResource
+
+
+class CustomsItem(CreateResource):
+    pass