+++ easypost/insurance.py
@@ -1,0 +1,5 @@
+from easypost.resource import AllResource, CreateResource
+
+
+class Insurance(AllResource, CreateResource):
+    pass