+++ easypost/pickup_rate.py
@@ -1,0 +1,5 @@
+from easypost.resource import Resource
+
+
+class PickupRate(Resource):
+    pass