+++ easypost/tax_identifier.py
@@ -1,0 +1,5 @@
+from easypost.resource import Resource
+
+
+class TaxIdentifier(Resource):
+    pass